requirements.txt
sqlalchemy~=1.4.22
pandas~=1.3.4
scikit-learn~=1.1.1
catboost~=1.0.4
numpy~=1.20.3
plotly~=5.6.0
streamlit~=1.7.0
joblib~=1.1.0
pillow~=8.4.0
scikit-image~=0.18.3
mlflow~=1.26.0
hyperopt~=0.2.7
yellowbrick~=1.4
lightgbm~=3.3.2
pyod~=1.0.1